extract options to configUtils
