--- src/config.js
@@ -1,3 +1,0 @@
-import config from '@siteConfig/config.yaml'
-
-export default config
